+++ app/models/locations.rb
@@ -6,24 +6,26 @@ require_dependency 'locations_table'
 class Locations
   extend LocationsTable
 
-  Struct.new('LocationSearchType', :label, :rdf_types)
+  # Prevent a warning about LocationSearchType being redefined
+  LocationSearchType = Struct.new('LocationSearchType', :label, :rdf_types)
+
+  # Formatting update to reduce line length warnings from Rubocop
+  EUROPEAN_REGION = ['http://data.ordnancesurvey.co.uk/ontology/admingeo/EuropeanRegion'].freeze
+  LOCAL_AUTHORITIES = [
+    'http://data.ordnancesurvey.co.uk/ontology/admingeo/Borough',
+    'http://data.ordnancesurvey.co.uk/ontology/admingeo/District',
+    'http://data.ordnancesurvey.co.uk/ontology/admingeo/LondonBorough',
+    'http://data.ordnancesurvey.co.uk/ontology/admingeo/MetropolitanDistrict',
+    'http://data.ordnancesurvey.co.uk/ontology/admingeo/UnitaryAuthority'
+  ].freeze
+  ENGLAND_REGION = ['http://landregistry.data.gov.uk/def/ukhpi/Region'].freeze
+  ENGLAND_COUNTY = ['http://data.ordnancesurvey.co.uk/ontology/admingeo/County'].freeze
+
   LOCATION_SEARCH_TYPES = {
-    'country' => Struct::LocationSearchType.new('Country', [
-                                                  'http://data.ordnancesurvey.co.uk/ontology/admingeo/EuropeanRegion'
-                                                ]),
-    'local_authority' => Struct::LocationSearchType.new('Local authority', [
-                                                          'http://data.ordnancesurvey.co.uk/ontology/admingeo/Borough',
-                                                          'http://data.ordnancesurvey.co.uk/ontology/admingeo/District',
-                                                          'http://data.ordnancesurvey.co.uk/ontology/admingeo/LondonBorough',
-                                                          'http://data.ordnancesurvey.co.uk/ontology/admingeo/MetropolitanDistrict',
-                                                          'http://data.ordnancesurvey.co.uk/ontology/admingeo/UnitaryAuthority'
-                                                        ]),
-    'england_region' => Struct::LocationSearchType.new('Region (England only)', [
-                                                         'http://landregistry.data.gov.uk/def/ukhpi/Region'
-                                                       ]),
-    'england_county' => Struct::LocationSearchType.new('County (England only)', [
-                                                         'http://data.ordnancesurvey.co.uk/ontology/admingeo/County'
-                                                       ])
+    'country' => LocationSearchType.new('Country', EUROPEAN_REGION),
+    'local_authority' => LocationSearchType.new('Local authority', LOCAL_AUTHORITIES),
+    'england_region' => LocationSearchType.new('Region (England only)', ENGLAND_REGION),
+    'england_county' => LocationSearchType.new('County (England only)', ENGLAND_COUNTY)
   }.freeze
 
   # @return The Location with the given URI, if it exists